LPS-101720 Rename *.checker.email.* to *.email.*
