+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-email-otp-api/bnd.bnd
@@ -1,8 +1,8 @@
 Bundle-Name: Liferay Multi-Factor Authentication Checker Email OTP API
-Bundle-SymbolicName: com.liferay.multi.factor.authentication.checker.email.otp.api
+Bundle-SymbolicName: com.liferay.multi.factor.authentication.email.otp.api
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.multi.factor.authentication.checker.email.otp.exception,\
-	com.liferay.multi.factor.authentication.checker.email.otp.model,\
-	com.liferay.multi.factor.authentication.checker.email.otp.service,\
-	com.liferay.multi.factor.authentication.checker.email.otp.service.persistence
+	com.liferay.multi.factor.authentication.email.otp.exception,\
+	com.liferay.multi.factor.authentication.email.otp.model,\
+	com.liferay.multi.factor.authentication.email.otp.service,\
+	com.liferay.multi.factor.authentication.email.otp.service.persistence
