+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-email-otp-api/src/main/java/com/liferay/multi/factor/authentication/checker/email/otp/model/MFAEmailOTPEntrySoap.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.multi.factor.authentication.checker.email.otp.model;
+package com.liferay.multi.factor.authentication.email.otp.model;
 
 import java.io.Serializable;
