+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-email-otp-service/bnd.bnd
@@ -1,5 +1,5 @@
 Bundle-Name: Liferay Multi-Factor Authentication Checker Email OTP Service
-Bundle-SymbolicName: com.liferay.multi.factor.authentication.checker.email.otp.service
+Bundle-SymbolicName: com.liferay.multi.factor.authentication.email.otp.service
 Bundle-Version: 1.0.0
 Liferay-Require-SchemaVersion: 1.0.0
 Liferay-Service: true
