+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-email-otp-service/service.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <!DOCTYPE service-builder PUBLIC "-//Liferay//DTD Service Builder 7.3.0//EN" "http://www.liferay.com/dtd/liferay-service-builder_7_3_0.dtd">
 
-<service-builder auto-import-default-references="false" auto-namespace-tables="false" mvcc-enabled="true" package-path="com.liferay.multi.factor.authentication.checker.email.otp">
+<service-builder auto-import-default-references="false" auto-namespace-tables="false" mvcc-enabled="true" package-path="com.liferay.multi.factor.authentication.email.otp">
 	<author>Arthur Chan</author>
 	<namespace>MFAEmailOTP</namespace>
 	<entity local-service="true" name="MFAEmailOTPEntry" remote-service="false">
