+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-email-otp-service/src/main/java/com/liferay/multi/factor/authentication/checker/email/otp/model/impl/MFAEmailOTPEntryImpl.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.multi.factor.authentication.checker.email.otp.model.impl;
+package com.liferay.multi.factor.authentication.email.otp.model.impl;
 
 /**
  * @author Arthur Chan
