+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-email-otp-service/src/main/resources/META-INF/portlet-model-hints.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 
 <model-hints>
-	<model name="com.liferay.multi.factor.authentication.checker.email.otp.model.MFAEmailOTPEntry">
+	<model name="com.liferay.multi.factor.authentication.email.otp.model.MFAEmailOTPEntry">
 		<field name="mvccVersion" type="long" />
 		<field name="mfaEmailOTPEntryId" type="long" />
 		<field name="companyId" type="long" />
