+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-email-otp-web/bnd.bnd
@@ -1,4 +1,4 @@
 Bundle-Name: Liferay Multi-Factor Authentication Checker Email OTP Web
-Bundle-SymbolicName: com.liferay.multi.factor.authentication.checker.email.otp.web
+Bundle-SymbolicName: com.liferay.multi.factor.authentication.email.otp.web
 Bundle-Version: 1.0.0
-Web-ContextPath: /multi-factor-authentication-checker-email-otp-web
+Web-ContextPath: /multi.factor.authentication.email-otp-web
