+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-email-otp-web/src/main/resources/META-INF/resources/init.jsp
@@ -20,7 +20,7 @@
 taglib uri="http://liferay.com/tld/frontend" prefix="liferay-frontend" %><%@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %>
 
-<%@ page import="com.liferay.multi.factor.authentication.checker.email.otp.web.internal.configuration.EmailOTPConfiguration" %><%@
+<%@ page import="com.liferay.multi.factor.authentication.email.otp.web.internal.configuration.EmailOTPConfiguration" %><%@
 page import="com.liferay.portal.kernel.module.configuration.ConfigurationProviderUtil" %><%@
 page import="com.liferay.portal.kernel.settings.LocalizedValuesMap" %>
